Porting watcher.js to typescript.
