+++ package.json
@@ -40,6 +40,7 @@
     "typescript": "^3.8.3"
   },
   "scripts": {
+    "build": "tsc",
     "test": "mocha --exit",
     "send-coverage": "node_modules/.bin/codeclimate-test-reporter < coverage/lcov.info; exit 0",
     "lint": "node_modules/.bin/eslint .",
